Add patchBox2DMath
