+++ patch/index.d.ts
@@ -2,7 +2,7 @@ import * as b2 from "box2dweb";
 import * as Box2D from "../lib";
 
 /**
- * パッチのオプショナルなパラメータ。
+ * patchBox2D() のオプショナルなパラメータ。
  */
 export interface PatchBox2DOptions {
     /**
@@ -23,3 +23,44 @@ export interface PatchBox2DOptions {
  * @param opts オプション。
  */
 export function patchBox2D(box2d: Box2D.Box2D, opts?: PatchBox2DOptions);
+
+
+/**
+ * patchBox2DMath() のオプショナルなパラメータ。
+ */
+export interface PatchBox2DMathOption {
+    /**
+     * 関数テーブルのサイズ。
+     * 省略時、 32768。
+     */
+    tableSize?: number;
+
+    /**
+     * 関数テーブルに収められる三角関数の区間。
+     *
+     * 真のとき、 [0, PI * 2] 。
+     * 偽のとき、 [0, PI / 2] 。
+     * 省略時、 true 。
+     *
+     * 区間を小さくするとことで精度が向上しますが、テーブルアクセスのコストが増えます。
+     */
+    wholePeriod?: boolean;
+
+    /**
+     * テーブルを格納する配列の型。
+     *
+     * 省略時、 Float32Array 。 Float32Array が利用できないとき、 Array 。
+     */
+    arrayType?: Float32Array | Float64Array | Array;
+}
+
+/**
+ * Box2Dの利用する三角関数にパッチをあてる。
+ *
+ * 以下の修正・機能追加を行う。
+ * - ブラウザ間の三角関数の計算結果の違いをなくするため、三角関数テーブルを用いる。
+ *
+ * @param box2d Box2D。
+ * @param opts オプション。
+ */
+export function patchBox2DMath(opts?: PatchBox2DMathOption);
